sepolicy_vndr: Move qti-media file contexts to genfs
[1] sku version file contexts are loaded by early-
init phase and sku version node is created after
file contexts are loaded. The transistion time
from early init to post-fs is around ~2.5 seconds.
avc denied messages are observed when sku_version
is accessed as vendor file contexts are slow in
reporting.

[2] Hence, move file contexts to genfs as genfs
helps to track and label the node even if the
node is created afterwards.

Change-Id: I97f9623cae1961accdd3b41e404bdf530c03f37b